avoid email loops by dropping suspect messages into the bit bucket
